# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlixQuest Scraper API",
    "version": "2.0.0",
    "description": "Exhaustive OpenAPI specification for FlixQuest Scraper API v2, designed for integration with Flutter applications."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
    "/api/v2/providers": {
      "get": {
        "summary": "List Available Providers",
        "description": "Retrieves the list of active (and optionally disabled) video scraper providers.",
        "operationId": "getProviders",
        "tags": ["Providers"],
        "parameters": [
          {
            "name": "all",
            "in": "query",
            "required": false,
            "description": "Set to 'true' to include disabled providers in the response.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "includeDisabled",
            "in": "query",
            "required": false,
            "description": "Alias for 'all'. Set to 'true' to include disabled providers.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "fProxy",
            "in": "query",
            "required": false,
            "description": "Forward proxy configuration mode or proxy URL.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider list retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/providers/{id}": {
      "patch": {
        "summary": "Update Provider Enabled State",
        "description": "Enables or disables a specific scraper provider by ID.",
        "operationId": "updateProviderStatus",
        "tags": ["Providers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique provider identifier (e.g. 'vixsrc').",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider status successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProviderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Provider not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/providers/{id}/toggle": {
      "post": {
        "summary": "Toggle Provider Status",
        "description": "Toggles the enabled status of a provider, or sets it explicitly if provided in the body.",
        "operationId": "toggleProviderStatus",
        "tags": ["Providers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique provider identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleProviderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider status successfully toggled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProviderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Provider not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/stream-movie": {
      "get": {
        "summary": "Get Movie Stream Links",
        "description": "Fetches validated playback stream links for a movie by TMDB ID using a selected provider.",
        "operationId": "getMovieStream",
        "tags": ["Streams"],
        "parameters": [
          {
            "name": "tmdbId",
            "in": "query",
            "required": true,
            "description": "The TMDB ID of the movie (e.g. '556574').",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "The provider identifier to scrape streams from (e.g. 'vixsrc').",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "proxy",
            "in": "query",
            "required": false,
            "description": "Set to 'false' or '0' to disable proxying and retrieve raw URLs when possible.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "noProxy",
            "in": "query",
            "required": false,
            "description": "Set to 'true' or '1' to bypass stream URL proxying.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fProxy",
            "in": "query",
            "required": false,
            "description": "Forward proxy setting (boolean or HTTP proxy string).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stream links retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid tmdbId or provider parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Provider not found, disabled, or no streams found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server or scraping error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/stream-tv": {
      "get": {
        "summary": "Get TV Episode Stream Links",
        "description": "Fetches validated playback stream links for a TV episode by TMDB ID, season, and episode using a selected provider.",
        "operationId": "getTVStream",
        "tags": ["Streams"],
        "parameters": [
          {
            "name": "tmdbId",
            "in": "query",
            "required": true,
            "description": "The TMDB ID of the TV show (e.g. '2316').",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": true,
            "description": "Season number (positive integer).",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "episode",
            "in": "query",
            "required": true,
            "description": "Episode number (positive integer).",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "The provider identifier to scrape streams from (e.g. 'vixsrc').",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "proxy",
            "in": "query",
            "required": false,
            "description": "Set to 'false' or '0' to disable stream proxying.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "noProxy",
            "in": "query",
            "required": false,
            "description": "Set to 'true' or '1' to bypass stream URL proxying.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fProxy",
            "in": "query",
            "required": false,
            "description": "Forward proxy setting (boolean or HTTP proxy string).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stream links retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Provider not found, disabled, or no streams found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server or scraping error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/dlhd/channels": {
      "get": {
        "summary": "List DLHD Live Channels",
        "description": "Scrapes the current 24/7 channel directory from DLHD.",
        "operationId": "getDlhdChannels",
        "tags": ["DLHD"],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Case-insensitive channel name or ID filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "description": "Set to true to bypass the channel cache.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current DLHD channels.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "source", "count", "channels"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "source": {
                      "type": "string",
                      "example": "dlhd"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "channels": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DlhdChannel"
                      }
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "DLHD channel directory was unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/dlhd/channels/{id}/stream": {
      "get": {
        "summary": "Extract a DLHD Channel HLS Stream",
        "description": "Resolves a numeric DLHD channel ID to a validated, short-lived M3U8 URL and the Origin, Referer, and User-Agent headers required for playback.",
        "operationId": "getDlhdChannelStream",
        "tags": ["DLHD"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric DLHD channel ID.",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "example": "51"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fresh HLS URL and playback headers. The response is not cacheable.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "source",
                    "channel",
                    "stream",
                    "extractedAt"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "source": {
                      "type": "string",
                      "example": "dlhd"
                    },
                    "channel": {
                      "$ref": "#/components/schemas/DlhdChannel"
                    },
                    "stream": {
                      "$ref": "#/components/schemas/DlhdStream"
                    },
                    "extractedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The channel ID is not numeric.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "No current playable M3U8 could be extracted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProviderListResponse": {
        "type": "object",
        "required": ["success", "providers"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderInfo"
            }
          }
        }
      },
      "UpdateProviderRequest": {
        "type": "object",
        "required": ["enabled"],
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "True to enable provider, false to disable.",
            "example": true
          }
        }
      },
      "ToggleProviderRequest": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Optional boolean to set status. If omitted, status will be toggled.",
            "example": true
          }
        }
      },
      "UpdateProviderResponse": {
        "type": "object",
        "required": ["success", "message", "provider"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Provider 'vixsrc' has been enabled"
          },
          "provider": {
            "$ref": "#/components/schemas/ProviderInfo"
          }
        }
      },
...
      "ProviderResponse": {
        "type": "object",
        "required": ["success", "provider"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "provider": {
            "type": "string",
            "example": "vixsrc"
          },
          "media": {
            "$ref": "#/components/schemas/MediaInfo"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderLink"
            }
          },
          "error": {
            "type": "string"
          },
          "details": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Missing or invalid tmdbId parameter"
          },
          "details": {
            "type": "string",
            "example": "Available providers: vixsrc, vidsrc, notorrent, ..."
          }
        }
      }
    }
  }
}